Main: Key-Points and Descriptors Compare
This is the main program. Here, the query image is taken from the user and the key-points and descriptors are extracted of that image and are compared with that of object images.
And if enough number of good matches are found, then the name of the object is fetched from the data set and is displayed to the user.